+++ pczt/src/roles/creator/mod.rs
@@ -123,7 +123,8 @@ impl Creator {
             zcash_primitives::transaction::TxVersion::V4 => Some(V4_TX_VERSION),
             zcash_primitives::transaction::TxVersion::V5 => Some(V5_TX_VERSION),
             #[cfg(zcash_unstable = "nu7")]
-            zcash_primitives::transaction::TxVersion::V6 => Some(V6_TX_VERSION),
+            zcash_primitives::transaction::TxVersion::V6
+            | zcash_primitives::transaction::TxVersion::V6_Qr => Some(V6_TX_VERSION),
             #[cfg(zcash_unstable = "zfuture")]
             zcash_primitives::transaction::TxVersion::ZFuture => None,
         }?;