+++ zcash_primitives/src/transaction/sighash.rs
@@ -63,7 +63,7 @@ pub fn signature_hash<
         TxVersion::V5 => v5_signature_hash(tx, signable_input, txid_parts),
 
         #[cfg(zcash_unstable = "nu7")]
-        TxVersion::V6 => v6_signature_hash(tx, signable_input, txid_parts),
+        TxVersion::V6 | TxVersion::V6_Qr => v6_signature_hash(tx, signable_input, txid_parts),
         #[cfg(zcash_unstable = "zfuture")]
         TxVersion::ZFuture => v6_signature_hash(tx, signable_input, txid_parts),
     })